PNPG-156 Moved the new mobile dialog in a own component
