+++ src/pages/dashboardUsers/UsersPage/UsersPage.tsx
@@ -1,17 +1,4 @@
-import {
-  Grid,
-  Tab,
-  Tabs,
-  Button,
-  styled,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  Slide,
-  Typography,
-  Stack,
-  Box,
-} from '@mui/material';
+import { Grid, Tab, Tabs, Button, Stack } from '@mui/material';
 import TitleBox from '@pagopa/selfcare-common-frontend/components/TitleBox';
 import { useEffect, useMemo, useState } from 'react';
 import { trackEvent } from '@pagopa/selfcare-common-frontend/services/analyticsService';
@@ -19,9 +6,6 @@ import { resolvePathVariables } from '@pagopa/selfcare-common-frontend/utils/rou
 import { useTranslation, Trans } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import { useUnloadEventOnExit } from '@pagopa/selfcare-common-frontend/hooks/useUnloadEventInterceptor';
-import { TransitionProps } from '@mui/material/transitions';
-import React from 'react';
-import CloseIcon from '@mui/icons-material/Close';
 import { ButtonNaked } from '@pagopa/mui-italia';
 import { Product, ProductsMap } from '../../../model/Product';
 import { Party } from '../../../model/Party';
@@ -33,49 +17,7 @@ import UserTableNoData from '../components/UserTableNoData';
 import { ProductsRolesMap } from '../../../model/ProductRole';
 import { ENV } from '../../../utils/env';
 import { useIsMobile } from '../../../hooks/useIsMobile';
-
-const MobileDialog = styled(Dialog)(() => ({
-  '& .MuiDialog-container': {
-    height: 'auto',
-    bottom: 0,
-    position: 'absolute',
-    width: '100%',
-  },
-  '& .MuiPaper-root': {
-    borderRadius: '24px 24px 0px 0px',
-    width: '100%',
-    maxWidth: '100vw',
-    margin: 0,
-  },
-  '& .MuiDialogContent-root': {
-    paddingTop: '20px !important',
-    width: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    flex: '1 1 auto',
-    overflowX: 'hidden',
-  },
-  '& .MuiDialogActions-root': {
-    display: 'block',
-    textAlign: 'center',
-    padding: '20px 24px',
-
-    '.MuiButton-root': {
-      width: '100%',
-      margin: '10px 0',
-    },
-  },
-}));
-
-const Transition = React.forwardRef(function Transition(
-  props: TransitionProps & {
-    children: React.ReactElement;
-  },
-  ref: React.Ref<unknown>
-) {
-  return <Slide direction="up" ref={ref} {...props} />;
-});
+import MobileFilter from '../components/MobileFilter';
 
 interface Props {
   party: Party;
@@ -136,14 +78,6 @@ function UsersPage({ party, activeProducts, productsMap, productsRolesMap }: Pro
   }, [productsFetchStatus]);
 
   useEffect(() => trackEvent('USER_LIST', { party_id: party.partyId }), [party]);
-
-  useEffect(() => {
-    setOpenDialogMobile(false);
-  }, [!isMobile]);
-
-  const handleClose = () => {
-    setOpenDialogMobile(false);
-  };
 
   const setSelectedProductSection = (productId?: string) =>
     // eslint-disable-next-line functional/immutable-data
@@ -223,55 +157,17 @@ function UsersPage({ party, activeProducts, productsMap, productsRolesMap }: Pro
             </Stack>
           </Grid>
         </Grid>
-        <MobileDialog
-          open={openDialogMobile}
-          fullWidth
-          sx={{ alignItems: 'center' }}
-          onClose={handleClose}
-          TransitionComponent={Transition}
-        >
-          <DialogTitle p={3}>
-            <Grid container alignItems="center" justifyContent="space-between">
-              <Grid item>
-                <Typography
-                  variant="h4"
-                  sx={{ fontSize: '24px', fontWeight: 'fontWeightBold', textAlign: 'left' }}
-                >
-                  {t('usersTable.filterRole.addFilters')}
-                </Typography>
-              </Grid>
-              <Grid item>
-                <CloseIcon
-                  onClick={handleClose}
-                  sx={{
-                    color: 'action.active',
-                    width: '32px',
-                    height: '32px',
-                  }}
-                />
-              </Grid>
-            </Grid>
-          </DialogTitle>
-          <DialogContent sx={{ display: 'flex' }}>
-            <Box sx={{ flexGrow: 1, width: 'calc(100vw - 100px)' }}>
-              <UsersTableActions
-                disableFilters={loading}
-                loading={loading}
-                party={party}
-                products={activeProducts}
-                productsRolesMap={
-                  !selectedProductSection
-                    ? productsRolesMap
-                    : { [selectedProductSection]: productsRolesMap[selectedProductSection] }
-                }
-                filters={filters}
-                onFiltersChange={setFilters}
-                showSelcRoleGrouped={!selectedProductSection}
-                setOpenDialogMobile={setOpenDialogMobile}
-              />
-            </Box>
-          </DialogContent>
-        </MobileDialog>
+        <MobileFilter
+          loading={loading}
+          activeProducts={activeProducts}
+          filters={filters}
+          openDialogMobile={openDialogMobile}
+          party={party}
+          productsRolesMap={productsRolesMap}
+          selectedProductSection={selectedProductSection}
+          setFilters={setFilters}
+          setOpenDialogMobile={setOpenDialogMobile}
+        />
         {isMobile ? (
           <Grid item mt={isMobile ? 3 : 0}>
             <ButtonNaked color="primary" onClick={() => setOpenDialogMobile(true)}>
@@ -291,7 +187,7 @@ function UsersPage({ party, activeProducts, productsMap, productsRolesMap }: Pro
             }
             filters={filters}
             onFiltersChange={setFilters}
-            showSelcRoleGrouped={!selectedProductSection}
+            showSelcRoleGrouped={isPnpg ? false : !selectedProductSection}
             setOpenDialogMobile={setOpenDialogMobile}
           />
         )}
